Remove OpenMP threading (except for within Eigen)
However, it does not appear that Eigen uses OpenMP to accelerate
the operations we are performing anyway. Only a single core is
used now. The output writing is nto really a bottleneck at all
at this stage so we take the simple approach and do that in the
same thread as the calculations.