Introduce zlib to optionally compress the preprocessed data
At present this is done in a simplistic but slow way where we
allocate a buffer, comrpess the data into it, write it to disk,
then delete the buffer for each column.

Since each column to be compressed has the same initial size we
can improve upon this by recycling a single buffer for all
columns and avoiding the calls to new/delete. That should speed
things up quite a bit. This will come in a following commit.