Rename member variables with m_ prefix
This is not just stylistic. It is a preparatory step to ready the
way for when we need to take local copies of the working variables
to allow multiple columns to be processed in parallel as part
of the asynchronous Gibbs work.

With the prefix we will be able to easily tell if a variable is
local or global (object scope level).