Refactor m_epsilon and m_y_tilde
These vectors are too large to take copies of, so take a unique lock of
m_mutex at the beginning to update m_y_tilde and calculate it's dot
project. The dot product is used throughout. At the end of the function
to a unique lock of m_mutex and update m_epsilon. However, the value of
m_y_tilde may have been changed by another thread during this interval.