Update notes
m_y_tilde is only used to assist the computation. It gets updated at the
beginning of BayesRRmz::processColumn* and the values are used
throughout, but it is not used outside of this method.

m_dist needs to be protected somehow. It is not safe to simultaneously
generate distributions on different threads using the same generator.