Take a local copy of epsilon and reduce work whilst holding lock
Taking a local copy of epsilon allows us to move more calculations
outside the scope where we hold either the shared lock or the
unique lock. This reduces the time per iteration quite a bit here.

Also, change to use std::memcpy for copying the large vectors
around. The operator = in Eigen uses a loop over pairs of doubles
and then a SIMD intrinsic to copy each pair. This is fine in the
common single threaded case for lazy evaluation of expression
templates but hurts us here when we want to minimise the amount
of work done whilst holding the lock.