Reduce the parallel job count to 6
This is the number of physical cores here. It seems that logical
hyperthread cores are nowhere near as good as real physical cores
when they are loaded with this kind of work.

Reducing this parallel job count massively speeds things up here
to the tune of 0.9s per iteration with the

ukb_imp_v3_UKB_EST_clumpLD09_chr22

dataset. I'm guessing that this is due to needing to shift data
between caches less often.