Don't copy large data between graph nodes
BayesRBase::processColumnAsync was returning a VectorXd of size equal to
the number of individuals by value. This vector was then copied between
nodes in ParallelGraph. Messages passed between nodes should be as light
weight as possible.

BayesRBase::processColumnAsync now returns a
std::unique_ptr<AsyncResult> which encapsulates the result of the
analysis and allows us to pass a pointer to the data between nodes.