Token based ParallelGraph
ParallelGraph now uses tokens to control the number of decompressed
markers stored in memory and the number of asynchronous analyses run
before synchronisation is forced.

Concurrency of decompression and analysis nodes can be controlled
independently, but the default is to let TBB manage this. Instead, the
synchronicity of the algorithm should be managed using tokens.

The following commandline options have been added
--decompression-concurrency <n> The maximum number of decompression_node
bodies which can run concurrently. 0 indicates unlimited concurrency.
The default value is 0.

--decompression-tokens <n> The maximum number of decompressed markers in
memory. Default value is 40.

--analysis-concurrency <n> The maximum number of analysis_node bodies
which can run concurrently. 0 indicates unlimited concurrency. The
default value is 0.

--analysis-tokens <n> The number of asynchronous analyses to do before
forcing synchronisation. Default value is 20.