Don't set limit_node threshold to zero
Setting the limit_node threshold to zero prevents any messages from
traversing the graph.

When --thread is not provided, brr defaults to maxParallel to 0 which
matches tbb::concurency::unlimited. This is fine for function_nodes
but not limit_nodes. Instead, we set the limit_threshold to the
maximum concurrency for this task arena.