Separate Marker into Kernel and Marker
The Marker struct only encapsulates the data for a single SNP. This
includes the code for reading and writing it to disk.

The Kernel struct encapsulates the data and algorithms required to
perform a particular analysis type on an individual marker.

This change allows us to reuse the marker data types with multiple
algorithm types.