Add basic MarkerCache
This MarkerCache loads all the Markers on disk and caches them into RAM.
It provides a very simple API which is not suited to populating the
cache in parallel.

This provides us with a basic cache to test performance implication of
caching Markers on the ParallelGraph.

The API is likely to change to support concurrent cache population in
the future.