Fix locking of m_epsilon in Bayes* implementations
The previous understanding of the flow graph was that the global update
node (calling Analysis::updateGlobal) would not run whilst the analysis
node (calling Analysis::processColumnAsync) was running. Whilst this is
correct for a message flowing through the ParallelGraph, it is not
correct for all threads.

Now we ensure that updates to m_epsilon are protected across threads.