Analysis::rubGibbs can be given a list of indexes
The list contains the index of each SNP to use in this analysis. This
allows us to run an analysis using only a subset of the markers in the
data set.

This will be useful for the hybrid mpi implementation, where each
process will work on a subset of markers and share results with the
other processes.