Add MPI API to Analysis
Analysis::accumulate should be called following an
Analysis::updateGlobal call, with the same arguments. It allows the
Analysis to accumulate updates that will be shared with the other
processes at the next synchronisation point.

Analysis::updateMpi should be called to synchronise results between
processes. The intention here is to call MPI_Allgather for each
parameter that needs sharing between processes.

This commit contains an outline of the intended implementation for the
BayesRBase class, but does not contain any MPI calls. These will be
added in a follow up commit, along with implementations for the
BayesWBase class.