brr can now --split preprocessed data files.
Example usage:
brr
--split /tmp/split
--hybrid-mpi 1
--data-file ./data/uk10k_chr1_1mb/uk10k_chr1_1mb.bed
--pheno ./data/uk10k_chr1_1mb/test.phen
--sparse-data ragged

The --split <splitDestination> option writes a subset of the
preprocessed data file to the splitDestination directory.

If the --hybrid-mpi option is omitted, the preprocessed data file is
copied to the splitDestination.

The --hybrid-mpi option allows the algorithm to detect the MPI world
size and then split the preprocessed datafile into equal subsets for
each host. Each subset is written to the splitDestination which is
expected to be local to each host.

The --data-file argument is required to indicate the location of the
preprocessed data file, as with the other --anylysis-type arguments. It
does not need to exist for the split mode to operate, but the associated
metadata (fam, bim and phenotype files) must.

The --working-directory argument can also be supplied to redirect to
local preprocessed datafiles as with the other analysis types.

The preprocessed data type flag (--sparse-data ragged) is required to
ensure the correct file names are used.