Add new hybrid-MPI implementation
This implementation adds an additional step to ParallelGraph in MPI
mode.

Once a subset of markers has been processed by the graph, we check for
results that need to be communicated between hosts. If found, the
parameters used to calculate the results and communicated to other
hosts. The other hosts then compute the new results locally using an
inner flow graph. Each host then has equal values for epsilon, beta and
betasqnG.

The inner flow graph loads markers from disk, then uses the marker to
calculate the results using the communicated parameters.

This approach is much faster than communicating the results.