Add Data::updateIndividualsForBED/CSV methods
These methods remove NA individuals from the analysis, however the
preprocessed data is not correctly updated. This is will be fixed in a
follow-up commit.

Add tests to ensure the analyses still run correctly with removed
individuals. Note that the results are not correct yet due to incorrect
preprocessed data.