Use Data::indInfoVec to remove NA individuals from preprocessed data
We can no longer rely on Data::numInds because we now have: the total
number of individuals in the data file; and the number of active
individuals in analysis.

The number of active individuals is controlled by the fixed effects
file.

Remove the Data::numInds member so that it cannot be accidentally used.