Fix file splitter reporting
The wrong format specifier was being used to print the marker subset
range, resulting in incorrect values appearing in the command line
output. The marker subsets sizes being used were correct.

Update the PreprocessedFileSplitter output to make it clear which
markers are copied and what the marker subset used for the analysis will
be.